requirements.txt
httplib2==0.7.7
oauth2==1.5.211
python-dateutil==2.1
simplejson==3.1.0
six==1.2.0
wsgiref==0.1.2
-e git+git@github.com:bear/python-twitter.git#egg=twitter
-e git@github.com:BradWhittington/django-templated-email.git#egg=templated-email